requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bidict==0.22.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
deepl==1.14.0
dnspython==2.3.0
Flask==2.3.2
Flask-Login==0.6.2
Flask-PyMongo==2.3.0
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
PyJWT==2.7.0
pymongo==4.3.3
python-dotenv==1.0.0
python-engineio==4.4.1
python-socketio==5.8.0
pytz==2023.3
requests==2.31.0
SQLAlchemy==2.0.13
twilio==8.2.1
typing_extensions==4.5.0
urllib3==2.0.2
Werkzeug==2.3.4
WTForms==3.0.1
yarl==1.9.2